train loop backbone

+++ marionet/dataset/dataset.py
@@ -107,4 +107,4 @@ class MarioNetDataset(Dataset):
         :param image: Image in numpy format, C x H x W.
         """
         image = torch.tensor(image)
-        return self.transforms(image)
+        return self.transforms(image).float()
